Fix build errors from incomplete server refactor
- Remove duplicate run logic from runner.rs that was left over after
  server_run() was extracted to server/runner.rs
- Fix DataSource trait mismatch: collection IDs use i64, not String
- Remove premature cfg(native) feature gates (native feature not in this PR)
- Fix needless_borrow clippy warning in server/runner.rs
- Move spawn helper fns from runner_tests.rs to process_tests.rs where
  they are actually used; restore where.exe fix for Windows test
