+++ src/server/mod.rs
@@ -1,7 +1,5 @@
-// Server backend for Hegel.
-//
-// When the `native` feature is NOT enabled, this module provides the test
-// runner that communicates with a Python hegel-core server over Unix sockets.
+// Server backend for Hegel: communicates with a Python hegel-core server
+// over Unix sockets.
 
 pub(crate) mod data_source;
 pub(crate) mod process;
